Small fixes
*CompVoltage changed to 1.9
*Mask changed to be unsigned long
*Fix in stopwatch expression to solve issue when timer value is uint32 max

+++ team_src/can.c
@@ -6,7 +6,7 @@
  */
 #include "all.h"
 
-unsigned int mask;
+unsigned long mask;
 stopwatch_struct* can_watch;
 struct ECAN_REGS ECanaShadow;
